fix/forge-and-fabric (#551)
* refactor: Remove obsolete AllowUnencryptedTunnel from Connect Config

* feat: Enhance modded server support and documentation

fixes bug: Modern forwarding cause disconnect with NeoForge 1.21.1 #545
fixes #544

- Added a new guide for modded servers, detailing compatibility with Fabric and NeoForge, including setup instructions and required mods.
- Updated compatibility information in the existing compatibility guide to highlight support for modded servers.
- Introduced new links in the configuration for easy navigation to the modded servers guide.
- Improved VarInt handling in the protocol to ensure compatibility with modded setups, including tests for negative values and CrossStitch compatibility.
- Minor refactor in the Connect configuration for better readability.

* fix: Improve error handling in WriteVarInt for CrossStitch compatibility test

- Added error handling to the WriteVarInt function call in the CrossStitch compatibility test to ensure proper verification of byte encoding. This change enhances test reliability by providing clearer failure messages.

+++ pkg/edition/java/proto/util/reader.go
@@ -122,7 +122,7 @@ func ReadVarIntReturnN(r io.Reader) (result int, n int, err error) {
 				break
 			}
 		}
-		return int(val), n, nil
+		return int(int32(val)), n, nil
 	}
 
 	var bytesRead byte = 0
